@@ -4,26 +4,24 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.16.1 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 dependencies:
-  cupertino_icons: 1.0.2
+  cupertino_icons: 1.0.4
   flutter:
     sdk: flutter
   http: 0.13.4
   infinite_scroll_pagination: 3.1.0
   logger: 1.1.0
-  path: 1.8.0
-  path_provider: 2.0.9
+  path: 1.8.1
+  path_provider: 2.0.10
   provider: 6.0.2
   share: 2.0.4
   shimmer: 2.0.0
-  sqflite: 2.0.2
-  url_launcher: 6.0.20
+  sqflite: 2.0.2+1
+  url_launcher: 6.1.2
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
   lint: 1.8.2
 
 flutter:
